Main Menu and Pause Menu
Added a main menu and pause menu to the game. From the main menu, the player can start or quit the game, as well as adjust the difficulty of the ghosts that roam around. From the pause menu, the player can resume or quit the game, as well as return to the main menu.